uefi: add a guess_canonical_file_name_for_binary helper
This helper attempts to deduce the canonical file name (such as
BOOTX64.EFI, BOOTAA64.EFI, etc) for a given binary using the file
utility.

This is needed for cases when our binary file doesn't have the canonical
name initially, which would require the caller to precompute and create
a copy of said file otherwise.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>